+++ cpp/src/toppra/solver/qpOASES-wrapper.cpp
@@ -1,3 +1,4 @@
+#include <toppra/toppra.hpp>
 #include <toppra/solver/qpOASES-wrapper.hpp>
 
 #include <qpOASES.hpp>
@@ -52,6 +53,7 @@ bool qpOASESWrapper::solveStagewiseOptim(std::size_t i,
         const Bound& x, const Bound& xNext,
         Vector& solution)
 {
+  TOPPRA_LOG_DEBUG("stage: i="<<i);
   Eigen::Index N (nbStages());
   assert (i <= N);
 
@@ -100,6 +102,12 @@ bool qpOASESWrapper::solveStagewiseOptim(std::size_t i,
       h[1] = std::min(h[1], box.x[i][1]);
     }
   }
+
+  TOPPRA_LOG_DEBUG("lA: " << std::endl << m_lA);
+  TOPPRA_LOG_DEBUG("hA: " << std::endl << m_hA);
+  TOPPRA_LOG_DEBUG(" A: " << std::endl << m_A);
+  TOPPRA_LOG_DEBUG("l : " << std::endl << l);
+  TOPPRA_LOG_DEBUG("h : " << std::endl << h);
 
   qpOASES::returnValue res;
   // TODO I assumed 1000 is the argument nWSR of the SQProblem.init function.