+++ cpp/src/toppra/toppra.hpp
@@ -2,10 +2,18 @@
 #define TOPPRA_TOPPRA_HPP
 
 #include <memory>
+#include <iostream>
 
 #include <Eigen/Core>
 #include <Eigen/StdVector>
 #include <vector>
+
+#ifdef TOPPRA_DEBUG_ON
+#define TOPPRA_LOG_DEBUG(X) std::cout << "[DEBUG]: " <<  X << std::endl
+#else
+#define TOPPRA_LOG_DEBUG(X) if (true) {} else {std::cout << "[DEBUG]: " <<  X << std::endl;}
+#endif
+
 
 /// The TOPP-RA namespace
 namespace toppra {
@@ -39,6 +47,7 @@ namespace toppra {
   } // namespace constraint
 
   class Solver;
+  typedef std::shared_ptr<Solver> SolverPtr;
   namespace solver {
     class qpOASESWrapper;
   } // namespace solver