+++ HISTORY.md
@@ -2,19 +2,25 @@
 
 ## 0.2.4 (in progress)
 
-### Added
+## Added
 
 - [cpp] Initial cpp TOPPRA implementation.
 - [python] Implement a new trajectory class for specified velocities.
 
-### Changed
-- [cpp] Bug fixes in solver wrapper.
+## Changed
+
 - [python] Simplify TOPPRA class interface.
 - [python] [0e022c][cm-0e022c] Update README.md to reflect development roadmap.
 - [python] Format some source files with black.
 - [python] [#78][gh-78] Improve documentation structure.
 - [python] [#79][gh-79] Improve documentation structure.
 
+### Changed
+- [cpp] Bug fixes in solver wrapper.
+
+## Removed
+
+- Dropping support for Python2.7. Adding type annotation compatible to Python 3.
 
 ### 0.2.3 (Jan 25 2020)